Move calculator to flow directory - help-if-you-are-arrested-abroad
This calculator is specific to the help-if-you-are-arrested-abroad flow and so
should be co-located.

* SmartAnswer::Calculators:ArrestedAbroad
  -> SmartAnswer::HelpIfYouAreArrestedAbroadFlow::Calculator

* SmartAnswer::Calculators::ArrestedAbroadCalculatorTest
  -> SmartAnswer::HelpIfYouAreArrestedAbroadFlow::CalculatorTest

Since these changes do not affect the behaviour and all the regression tests
pass, I've updated the checksums with the new location of the calculator.
