+++ test/data/help-if-you-are-arrested-abroad-files.yml
@@ -1,5 +1,5 @@
 ---
-lib/smart_answer_flows/help-if-you-are-arrested-abroad.rb: c106f1d1ff22a0805eb59a710cf6d74b
+lib/smart_answer_flows/help-if-you-are-arrested-abroad.rb: c9e6f29324ab2a1315fc9195acc95fc5
 test/data/help-if-you-are-arrested-abroad-questions-and-responses.yml: c007b0dd259d1c09ccc2dd3163fa1e56
 test/data/help-if-you-are-arrested-abroad-responses-and-expected-results.yml: 3e3aeca8a1af7c147b8f62af365f650d
 lib/smart_answer_flows/help-if-you-are-arrested-abroad/help_if_you_are_arrested_abroad.govspeak.erb: cb923645274b287ba93af8c7118d7f88
@@ -9,5 +9,5 @@ lib/smart_answer_flows/help-if-you-are-arrested-abroad/outcomes/answer_one_gener
 lib/smart_answer_flows/help-if-you-are-arrested-abroad/outcomes/answer_three_syria.govspeak.erb: 3e703bedc21d8ec141a6bf4458ee8c94
 lib/smart_answer_flows/help-if-you-are-arrested-abroad/outcomes/answer_two_iran.govspeak.erb: 41d2d0b6f54f09d9543fac0e7e862b4d
 lib/smart_answer_flows/help-if-you-are-arrested-abroad/questions/which_country.govspeak.erb: 913b5637a2814cbd529affffc55b1f72
-lib/smart_answer/calculators/arrested_abroad.rb: 6b127f4dbdfd258cc9be5957c0f43352
+lib/smart_answer_flows/help-if-you-are-arrested-abroad/calculator.rb: 39db60e0870999aeb2c7142fe3ac579a
 lib/data/prisoner_packs.yml: 821fb0e9f893354f647d5b97453ce322
